+++ extensions/word-search/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Word Search Changelog
 
+## [Upgrades] - 2024-06-01
+
+- Updated dependencies to latest versions
+- Rewrote API using `useFetch` hook
+- Added option to use selection as search text ([#11748](https://github.com/raycast/extensions/issues/11748))
+- Added option to set the max results
+- Added search for words that sound like the search text
+- Added Word Suggestion command, including vocabulary switch (`English`, `Spanish` and `Wikipedia (English)`)
+- Added Details option for words
+
 ## [Output capitalized word] - 2023-02-28
 
 - Output will be capitalized when search text is capitalized